+++ extLibs/uil/src/main/java/com/nostra13/universalimageloader/Data.java
@@ -1,0 +1,8 @@
+package com.nostra13.universalimageloader;
+
+
+import sarsamurmu.adaptiveicon.AdaptiveIcon;
+
+public class Data {
+    public static int AdaptiveIconShape = AdaptiveIcon.PATH_CIRCLE;
+}